Move import-export to importExportStatic and remove live binding section
